Renamed Shared namespace to Common
